Improve compileOpFull API and custom op validation
API changes:
- Use std::span and std::tuple instead of flat C arrays for inputs/outputs
- Change shape/offset types from int to int64_t for PyTorch consistency
- Remove input/output sorting that broke user-specified tensor order

Validation improvements:
- Add dtype and shape validation for custom op inputs/outputs
- Include index, expected vs actual values in error messages

Other fixes:
- Flatten tensors in scatter/gather for compile_op 1D shape expectation
- Use IVector instead of std::vector for tensorShape fields
- Add Vector comparison operators for shape validation
- Fix file descriptor leak in tests/run.py (close child_conn after fork)

+++ core/processgroup.h
@@ -126,9 +126,9 @@ public:
 
   virtual void shutdown();
 
-  CustomOp compileOpFull(const std::vector<int>& shape, torch::Dtype dtype,
-      const std::vector<std::tuple<int, std::vector<int>, std::vector<int>>>& inputs,
-      const std::vector<std::tuple<int, std::vector<int>, std::vector<int>>>& outputs);
+  CustomOp compileOpFull(const std::vector<int64_t>& shape, torch::Dtype dtype,
+      const std::vector<std::tuple<int, std::vector<int64_t>, std::vector<int64_t>>>& inputs,
+      const std::vector<std::tuple<int, std::vector<int64_t>, std::vector<int64_t>>>& outputs);
 };
 
 } // namespace moodist
